Made a very small change to the Matlab interface for compatibility with Windows Visual Studio.
svn-id: 1635 @trunk

+++ Ipopt/contrib/MatlabInterface/src/matlabinfo.cpp
@@ -32,7 +32,7 @@ MatlabInfo::MatlabInfo (mxArray*& ptr)
 
 ApplicationReturnStatus MatlabInfo::getExitStatus() const {
   const mxArray* p = mxGetField(ptr,0,"status");
-  return (ApplicationReturnStatus) *mxGetPr(p);  
+  return (ApplicationReturnStatus) (int) *mxGetPr(p);  
 }
 
 void MatlabInfo::setExitStatus (ApplicationReturnStatus status) {
